Update BuilderCanvas.vue layout and enhance button properties
Adapted the BuilderCanvas.vue layout by modifying CSS classes. The vertical alignment of the items was changed from center to start to enhance visibility. For better UI, the height of 'Text' elements was increased and also of the 'Button' element. In addition, default 'backgroundColor' and 'color' properties were injected in the 'Button' element directly in the code for easier usage. All these changes contribute to an improved UX of the Builder Canvas.