Update BuilderCanvas to support two-way data binding
In 'BuilderCanvas.vue', the change was made from using ':value' directive to 'v-model' directive for binding the 'element.defaultProperties.text'. This was done to facilitate two-way data binding between textarea and the Vue component. With this change, any modifications to the text content of the textarea would automatically update 'element.defaultProperties.text' and vice versa, providing a more reactive and dynamic user experience.