Improve element manipulation and interaction
This revision brings considerable changes to the handling of elements within pages. We include an updatePage function to refresh the details of an existing page and modify the setCurrentPage function to optimize element assignment for the current page. To ensure consistency, modifications to canvas elements now trigger a corresponding update of elements on the current page. Element addition, removal, or reordering produce immediate and consistent results across pages and elements.

Comparable modifications are made in the BuilderCanvas.vue file to reflect the changes in the elements store. PageTab.vue and SideBar.vue now account for the modified interaction with elements.

The modifications also include additional SVG icons for mobile and desktop versions and logic to handle outside click events in App.vue. We refined ElementProperties.vue components for different elements, and a few more improvements were made to make the UI more user-friendly.

These changes significantly improve the user experience while interacting with components and pages, ensuring seamless and consistent changes across the application.