Update user interface styling and default button action
Enhancements were made to the UI styling of selected elements on 'BuilderCanvas.vue' for a clearer differentiation of selected elements. Button properties were updated in 'elements.ts', including setting default action to "Redirect to Next Page", and setting default margins to 0 for optimal initial layout. The UI in 'ElementProperties.vue' was also improved, changing the action input field to a dropdown for an overall improved user experience.