Improve selected element visualization in the BuilderCanvas
Developer added z-index property and an absolute div to the BuilderCanvas container in BuilderCanvas.vue, allowing the selected element to become identifiable in terms of depth, size and position. Fixed width property to minimum width(min-w) for better responsive design in PageTabs.vue. This changes improve user experience by making easier for users to visualize the hierarchy of their elements. Also, added an overflow CSS property to the App.vue and PageTabs.vue which prevents contents from spilling over the container, thus enhancing the overall look and usage of the application.