Add insert element action and update BuilderCanvas methods
Inserted an `insertElementToCanvas` action in `src/stores/elements.ts` to handle adding elements to Canvas at a specific index. This action creates a new element with a unique id and inserts it at a specified index in the `canvasElements` array. In `src/components/BuilderCanvas.vue`, the drop method is updated to use this action when a `dropIndex` is present. A new drap and drop functionality is introduced where `dropIndex` is updated in `dragOver` and reset in `drop` and `dragEnd`. It enhances the usability by allowing users to place elements at a specific location in the Canvas page.